# Improve API performance