--- a/src/aws_lambda_mpic/mpic_caa_checker_lambda/mpic_caa_checker_lambda_function.py
+++ b/src/aws_lambda_mpic/mpic_caa_checker_lambda/mpic_caa_checker_lambda_function.py
@@ -48,8 +48,10 @@ def get_handler() -> MpicCaaCheckerLambdaHandler:
         _handler = MpicCaaCheckerLambdaHandler()
     return _handler
 
-if os.environ.get("PYTEST_VERSION") is None:
+
+if os.environ.get("AWS_LAMBDA_FUNCTION_NAME") is not None:
     get_handler()
+
 
 # noinspection PyUnusedLocal
 # for now, we are not using context, but it is required by the lambda handler signature