… src/aws_lambda_mpic/__about__.py