@birgelee 2025-04-27
I spoke with @sciros and the AWS lambda name check is OK since we are in an AWS repo. I am going ahead with merging this in since all comments are addressed.